fix: Update Class 3 Intermediate certificate to allow chaining to re-signed Class 1 Root certificate
The Class 1 Root and Class 3 Intermediate certificates both contained an Authority Key Identifier extension,
that made it impossible to change the Class 1 Root certificate's serial. This lead to at least two known issues:

1. Presenting a browser with the re-signed root, with the same issuer/serial combination
causes the browser to reject that certificate.

2. Changing the Serial Number, but keeping with the AKI X.509v3 extension causes some
software (noticed with Firefox) to (completely by spec) to try building a chain that eventually
should reach the old Class 1 Root. This never happens, as although the re-signed Class 1 Root
is a valid signer of itself, it is not considered its issuer - causing an infinite chain to be built.

To allow the re-signed Class 1 Root to be considered the issuer of both itself and the
Class 3 Intermediate certificates the Authority Key Information should only contain the
Issuer Public Key Information part of the Authority Key Information, while skipping the
Directory Name and Serial details. We thus need to remove that part of the extension.